+++ app/bootstrap.php
@@ -3,6 +3,7 @@
 use Framework\App\App;
 use Framework\Kernel\HttpKernel;
 use Framework\Router\Router;
+use Framework\TemplateEngine\TwigTemplateEngine;
 
 
 // Instanciate DotEnv
@@ -22,9 +23,10 @@ $config = require_once __DIR__.'/config.php';
 // Create dependencies for the application
 $kernel = new HttpKernel();
 $router = new Router();
+$template_engine = new TwigTemplateEngine();
 
 // Create a new instance of the application
-$app = new App($config, $kernel, $router);
+$app = new App($config, $kernel, $router, $template_engine);
 
 // Return the instance to the caller script
 return $app;