+++ framework/App/AppInterface.php
@@ -5,6 +5,7 @@ namespace Framework\App;
 use Framework\Kernel\KernelInterface;
 use Framework\Request\RequestInterface;
 use Framework\Router\RouterInterface;
+use Framework\TemplateEngine\TemplateEngineInterface;
 
 interface AppInterface
 {
@@ -13,6 +14,7 @@ interface AppInterface
     public static function config() : Array;
     public static function kernel() : KernelInterface;
     public static function router() : RouterInterface;
+    public static function template_engine() : TemplateEngineInterface;
     public function set($key, $value);
     public function get($key);
     public function handle(RequestInterface $request) : void;