+++ framework/TemplateEngine/TemplateEngineInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Framework\TemplateEngine;
+
+use Framework\App\BootstrappableServiceInterface;
+
+interface TemplateEngineInterface extends BootstrappableServiceInterface
+{
+    public function render(String $template, Array $context=[]);
+}